# docs/swagger.yaml
basePath: /
definitions:
  erg_ninja_internal_dto_response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      path:
        type: string
      request_id:
        type: string
      statusCode:
        type: integer
      timestamp:
        type: string
    type: object
...
  erg_ninja_internal_modules_courses_dto.CourseResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      enrollmentCount:
        type: integer
      id:
        type: string
      instructorId:
        type: string
      ratingAvg:
        type: number
      schemaDataJson:
        type: string
      schemaType:
        type: string
      slug:
        type: string
      status:
        type: string
      subdomain:
        type: string
      tenantId:
        type: string
      themeConfig:
        type: string
      thumbnailUrl:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  erg_ninja_internal_modules_courses_dto.CreateCourseRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      instructorId:
        type: string
      schemaDataJson:
        type: string
      schemaType:
        description: Course, OnlineCourse
        type: string
      slug:
        maxLength: 100
        type: string
      subdomain:
        maxLength: 100
        type: string
      themeConfig:
        type: string
      thumbnailUrl:
        type: string
      title:
        maxLength: 200
        minLength: 3
        type: string
    required:
    - slug
    - title
    type: object
  erg_ninja_internal_modules_courses_dto.UpdateCourseRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      instructorId:
        type: string
      schemaDataJson:
        type: string
      schemaType:
        type: string
      slug:
        maxLength: 100
        type: string
      status:
        type: string
      subdomain:
        maxLength: 100
        type: string
      thumbnailUrl:
        type: string
      title:
        maxLength: 200
        minLength: 3
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: ERG Support
    url: https://erg.ninja/support
  description: The unified backend API for ERG — a multi-tenant eLearning, content
    management, and automation platform.
  title: ERG Platform API
  version: "1.0"
paths:
...
  /api/courses:
    get:
      consumes:
      - application/json
      description: Fetch a paginated list of courses with optional status and search
        filtering.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Filter by status (draft|published)
        in: query
        name: status
        type: string
      - description: Search in title or description
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/erg_ninja_internal_modules_courses_dto.CourseResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
      summary: List courses
      tags:
      - Courses
    post:
      consumes:
      - application/json
      description: Create a new course for the current tenant. Admin only.
      parameters:
      - description: Course Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/erg_ninja_internal_modules_courses_dto.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/erg_ninja_internal_modules_courses_dto.CourseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
      security:
      - BearerAuth: []
      summary: Create a new course
      tags:
      - Courses
  /api/courses/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a course and its associated lessons. Admin only.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
      security:
      - BearerAuth: []
      summary: Delete a course
      tags:
      - Courses
    get:
      consumes:
      - application/json
      description: Fetch full course details including lessons and settings.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erg_ninja_internal_modules_courses_dto.CourseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
      summary: Get course by ID
      tags:
      - Courses
    patch:
      consumes:
      - application/json
      description: Update course fields (title, description, status, etc.). Admin
        only.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/erg_ninja_internal_modules_courses_dto.UpdateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erg_ninja_internal_modules_courses_dto.CourseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
      security:
      - BearerAuth: []
      summary: Update an existing course
      tags:
      - Courses
  /api/courses/{id}/lessons/reorder:
    post:
      consumes:
      - application/json
      description: Reorders lessons within a course.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reorder lessons
      tags:
      - Courses
  /api/courses/{id}/schema:
    get:
      description: Returns JSON-LD schema.org markup for a course.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/ld+json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get course schema
      tags:
      - Courses
  /api/courses/{id}/theme:
    patch:
      consumes:
      - application/json
      description: Updates the visual theme configuration for a course.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update course theme
      tags:
      - Courses
  /api/courses/subdomain/{sub}:
    get:
      consumes:
      - application/json
      description: Fetch course details and lessons by the tenant's slug/subdomain.
      parameters:
      - description: Subdomain/Slug
        in: path
        name: sub
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/erg_ninja_internal_modules_courses_dto.CourseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/erg_ninja_internal_dto_response.Response'
      summary: Get course by subdomain
      tags:
      - Courses
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"